Retry dropping slot on errors
Postgres will error when attempting to drop a slot while it's actively
being used. This can present a problem when a single Terraform plan will
drop the replication slot as well as stopping the service using it and
thus make the slot inactive.In that scenario we'd like to keep trying to
drop the slot until the service has come down.

See [terraform docs][1] for how to override the default configuration of
retrying up to one minute.

[1]: https://www.terraform.io/docs/configuration/resources.html#operation-timeouts